# Slightly faster `HexEncoder.EncodeData` for .NET 6