@NicolasDorier 2022-12-07
Note that this PR because of not supporting lowercase is allocating way more.